requirements.txt
pillow==8.2.0
opencv-python==4.5.4.58